Simplify Data Dependencies By Separating Out Player History
In previous version, player opponent history was tracked in the player object. This made things messy when rescheduling by say adding a player because fixtures were dependent on updates to players, but the player opponent history is dependent on fixtures, hence a circular dependency of sorts. This will make things easier